requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
PyMuPDF==1.23.8
Pillow==10.0.1
reportlab==4.0.4
